fix(nostr): address remaining IronLoop review findings for PR #7184
1. Move relay I/O out of ironclaw_wasm lane into ironclaw_host_runtime
   - Relay subscribe/publish/close now live in host_runtime::nostr_relay
   - ironclaw_wasm delegates to host via WasmHostNostr trait (new methods)
   - WASM guests never open sockets directly; host mediates all relay I/O

2. Add per-scope Nostr capability gating (WasmRuntimeNostrAdapter)
   - host_for_scope wraps KernelNostrHost in scope/capability_id adapter
   - Network policy check now applies to Nostr relay calls
   - Audit logging includes scope + capability_id for every Nostr operation

3. Wire production Nostr host in composition layer
   - Added with_nostr_host() builder to HostRuntimeServices
   - KernelNostrHost made pub + re-exported from ironclaw_host_runtime
   - attach_nostr_host() reads IRONCLAW_REBORN_NOSTR_PRIVATE_KEY env var
   - Called after attach_wasm_runtime() in production backend assembly
   - Fail-closed: missing key = DenyWasmHostNostr, invalid key = warning + deny

Fixes: IronLoop review #2 (scope gating), #3 (relay I/O mediation), #5 (composition wiring)
