+++ crates/lanes/ironclaw_wasm/Cargo.toml
@@ -10,20 +10,14 @@ layer = "runtimes"
 [dependencies]
 ironclaw_host_api = { path = "../../contracts/ironclaw_host_api" }
 ironclaw_wasm_limiter = { path = "../ironclaw_wasm_limiter" }
-futures-util = "0.3"
 secp256k1 = { version = "0.29", features = ["rand-std", "global-context"] }
 serde_json = "1"
 bech32 = "0.12"
 sha2 = "0.10"
 thiserror = "2"
 tokio = { version = "1", features = ["rt", "rt-multi-thread", "time"] }
-# TLS is provided via rustls-tls-webpki-roots (uses rustls + webpki root certs).
-# This ensures all relay WebSocket connections use TLS with the standard
-# root certificate store — no additional TLS feature configuration needed.
-tokio-tungstenite = { version = "0.24", features = ["rustls-tls-webpki-roots"] }
 tracing = "0.1"
 url = "2"
-uuid = { version = "1", features = ["v4"] }
 wasmtime = { workspace = true }
 wasmtime-wasi = { workspace = true }
