+++ crates/lanes/ironclaw_wasm/src/lib.rs
@@ -24,10 +24,10 @@ pub use host::{
     SystemWasmHostClock, WasmHostClock, WasmHostHttp, WasmHostNostr, WasmHostSecrets,
     WasmHostTools, WasmHostWorkspace, WasmHttpRequest, WasmHttpResponse,
     WasmRuntimeCredentialProvider, WasmRuntimeCredentialRequest, WasmRuntimeHttpAdapter,
-    WasmRuntimePolicyDiscarder, WasmStagedRuntimeCredential, WasmStagedRuntimeCredentialScope,
+    WasmRuntimePolicyDiscarder, WasmRuntimeNostrAdapter, WasmStagedRuntimeCredential, WasmStagedRuntimeCredentialScope,
     WasmStagedRuntimeCredentials, WitToolHost,
 };
-pub use nostr_relay::{publish_nostr_event, subscribe_nostr_events, validate_relay_url};
+pub use nostr_relay::{validate_relay_url, NostrRelayError};
 pub use nostr_signer::{decode_nostr_private_key, sign_nostr_event, NostrSignError};
 pub use runtime::WitToolRuntime;
 pub use types::{PreparedWitTool, WasmLogLevel, WasmLogRecord, WitToolExecution, WitToolRequest};
